+++ src/core/__tests__/AtomId.test.ts
@@ -1,6 +1,6 @@
-import AtomId from '../core/AtomId';
-import IndexMap from '../core/IndexMap';
-import getAtomId from '../utils/factories/AtomId';
+import AtomId from '../AtomId';
+import IndexMap from '../IndexMap';
+import getAtomId from '../../utils/factories/AtomId';
 
 describe('Atom', () => {
   describe('toString', () => {
